Use separate auth service for spnego
