+++ cruise-control/src/test/java/com/linkedin/kafka/cruisecontrol/servlet/security/spnego/SpnegoUserStoreAuthorizationServiceTest.java
@@ -2,8 +2,10 @@
  * Copyright 2020 LinkedIn Corp. Licensed under the BSD 2-Clause License (the "License"). See License in the project root for license information.
  */
 
-package com.linkedin.kafka.cruisecontrol.servlet.security;
+package com.linkedin.kafka.cruisecontrol.servlet.security.spnego;
 
+import com.linkedin.kafka.cruisecontrol.servlet.security.DefaultRoleSecurityProvider;
+import com.linkedin.kafka.cruisecontrol.servlet.security.UserStoreAuthorizationService;
 import org.eclipse.jetty.security.UserStore;
 import org.eclipse.jetty.server.UserIdentity;
 import org.junit.Test;
@@ -11,7 +13,7 @@ import org.junit.Test;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
-public class UserStoreAuthorizationServiceTest {
+public class SpnegoUserStoreAuthorizationServiceTest {
 
   private static final String TEST_USER = "testUser";
 
@@ -19,7 +21,7 @@ public class UserStoreAuthorizationServiceTest {
   public void testPrincipalNames() {
     UserStore users = new UserStore();
     users.addUser(TEST_USER, null, new String[] { DefaultRoleSecurityProvider.ADMIN });
-    UserStoreAuthorizationService usas = new UserStoreAuthorizationService(users);
+    UserStoreAuthorizationService usas = new SpnegoUserStoreAuthorizationService(users);
 
     UserIdentity result = usas.getUserIdentity(null, TEST_USER + "/host@REALM");
     assertNotNull(result);
@@ -30,6 +32,10 @@ public class UserStoreAuthorizationServiceTest {
     assertEquals(TEST_USER, result.getUserPrincipal().getName());
 
     result = usas.getUserIdentity(null, TEST_USER + "/host");
+    assertNotNull(result);
+    assertEquals(TEST_USER, result.getUserPrincipal().getName());
+
+    result = usas.getUserIdentity(null, TEST_USER);
     assertNotNull(result);
     assertEquals(TEST_USER, result.getUserPrincipal().getName());
   }
